requirements.txt
# Web Framework
fastapi>=0.111
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2

# PDF Processing
pdfminer.six==20221105
PyPDF2==3.0.1

# Data Processing
pandas==2.1.1
numpy==1.26.0

# Document Generation
python-docx==1.0.1

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.2

# Utilities
python-dotenv==1.0.0

# API Clients
anthropic==0.49.0

# Authentication
boto3==1.28.62
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2

